TEA.java
Updated the passwordEncode and passwordDecode methods. The passwordEncode method now adds a unique salt after the password and places the password inside the passMask byte array. The password size should be 12 bytes max and the key size should be 4 bytes max. These need to be checked prior to using the class. All result byte arrays and the passMask array has been changed to 24 bytes. This convention comes from the client because the client sends the encrypted password in the size of 24 bytes.   Also, check to ensure the code runs on your version of Java. I'm using JDK14.